move SWAHparams
